01-models/Wan/Wan2.1-T2V-1.3B-Diffusers/requirements.txt
torchvision==0.21.0
opencv-python==4.11.0.86
diffusers==0.35.2
transformers==4.49.0
tokenizers==0.21.1
accelerate==1.4.0
peft==0.17.1
ftfy==6.3.1
ffmpeg==1.4
imageio==2.37.2
imageio-ffmpeg==0.6.0
